# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for Anki Backend system. This API provides endpoints for managing decks, notes, cards, and study sessions.",
        "title": "Anki Backend API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@anki.example.com"
        },
        "license": {
            "name": "MIT",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/decks": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns all decks belonging to the authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "decks"
                ],
                "summary": "List all decks",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/response.DeckResponse"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a new deck for the authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "decks"
                ],
                "summary": "Create a new deck",
                "parameters": [
                    {
                        "description": "Deck creation request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.CreateDeckRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/response.DeckResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/decks/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns a deck by its ID for the authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "decks"
                ],
                "summary": "Get deck by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Deck ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.DeckResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates an existing deck's name, parent or options",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "decks"
                ],
                "summary": "Update deck",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Deck ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Deck update request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.UpdateDeckRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.DeckResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Soft deletes a deck and its sub-decks",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "decks"
                ],
                "summary": "Delete deck",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Deck ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "request.CreateDeckRequest": {
            "description": "Request payload for creating a new deck",
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "description": "Name of the deck",
                    "type": "string",
                    "example": "Idiomas::Inglês"
                },
                "options_json": {
                    "description": "Configuration options for the deck (JSON string)",
                    "type": "string",
                    "example": "{}"
                },
                "parent_id": {
                    "description": "ID of the parent deck (optional)",
                    "type": "integer",
                    "example": 1
                }
            }
        },
...
        "request.UpdateDeckRequest": {
            "description": "Request payload for updating an existing deck",
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "description": "New name of the deck",
                    "type": "string",
                    "example": "Idiomas::Inglês::Avançado"
                },
                "options_json": {
                    "description": "New configuration options for the deck (JSON string)",
                    "type": "string",
                    "example": "{}"
                },
                "parent_id": {
                    "description": "New parent deck ID (optional)",
                    "type": "integer",
                    "example": 2
                }
            }
        },
...
        "response.DeckResponse": {
            "description": "Response payload containing deck information",
            "type": "object",
            "properties": {
                "created_at": {
                    "description": "Timestamp when the deck was created",
                    "type": "string",
                    "example": "2024-01-15T10:30:00Z"
                },
                "id": {
                    "description": "Unique identifier for the deck",
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "description": "Name of the deck",
                    "type": "string",
                    "example": "Idiomas::Inglês"
                },
                "options_json": {
                    "description": "Configuration options as a JSON string",
                    "type": "string",
                    "example": "{}"
                },
                "parent_id": {
                    "description": "ID of the parent deck (null if it's a root deck)",
                    "type": "integer"
                },
                "updated_at": {
                    "description": "Timestamp when the deck was last updated",
                    "type": "string",
                    "example": "2024-01-15T10:30:00Z"
                },
                "user_id": {
                    "description": "ID of the user who owns the deck",
                    "type": "integer",
                    "example": 1
                }
            }
        },
...
        "response.ErrorResponse": {
            "description": "Standardized error response format returned by the API",
            "type": "object",
            "properties": {
                "code": {
                    "description": "Error code (same as error field, kept for compatibility)",
                    "type": "string",
                    "example": "NOT_FOUND"
                },
                "error": {
                    "description": "Error code (e.g., \"NOT_FOUND\", \"VALIDATION_ERROR\")",
                    "type": "string",
                    "example": "NOT_FOUND"
                },
                "message": {
                    "description": "Human-readable error message",
                    "type": "string",
                    "example": "Resource not found"
                },
                "path": {
                    "description": "Request path",
                    "type": "string",
                    "example": "/api/users/123"
                },
                "request_id": {
                    "description": "Request ID for tracking",
                    "type": "string",
                    "example": "abc123def456"
                },
                "timestamp": {
                    "description": "ISO 8601 timestamp",
                    "type": "string",
                    "example": "2024-01-15T10:30:00Z"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token. Example: \"Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}